to_do_list/requirement.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
drf-yasg